chore: refactor and cleanup theme-ui styles
