--- theme/src/gatsby-plugin-theme-ui/abstracts/fonts.js
@@ -1,6 +1,0 @@
-export default {
-  sans: '-apple-system, BlinkMacSystemFont, avenir next, avenir, helvetica neue, helvetica, Ubuntu, roboto, noto, segoe ui, arial, sans-serif',
-  serif:
-    'Iowan Old Style, Apple Garamond, Baskerville, Times New Roman, Droid Serif, Times, Source Serif Pro, serif, Apple Color Emoji, Segoe UI Emoji, Segoe UI Symbol',
-  mono: 'Menlo, Consolas, Monaco, Liberation Mono, Lucida Console, monospace'
-}
