--- theme/src/gatsby-plugin-theme-ui/abstracts/shadows.js
@@ -1,8 +1,0 @@
-export const floatOnHover = {
-  transition: `all 200ms ease-in-out`,
-
-  '&:hover, &:focus': {
-    transform: `scale(1.015)`,
-    boxShadow: `lg`
-  }
-}
